+++ packages/scenes/src/variables/macros/AllVariablesMacro.test.ts
@@ -5,7 +5,8 @@ import { TestVariable } from '../variants/TestVariable';
 import { AllVariablesMacro } from './AllVariablesMacro';
 import { ALL_VARIABLE_TEXT, ALL_VARIABLE_VALUE } from '../constants';
 import { TestScene } from '../TestScene';
-import { VARIABLE_NAMESPACE } from '../../core/types';
+
+import { VARIABLE_NAMESPACE } from '../utils';
 
 describe('UrlVariables', () => {
   it('Should include variables from all levels', () => {
