+++ packages/scenes/src/variables/utils.ts
@@ -6,6 +6,12 @@ import { SceneDataQuery, SceneObject, SceneObjectState } from '../core/types';
 import { SceneQueryRunner } from '../querying/SceneQueryRunner';
 import { DataSourceRef } from '@grafana/schema';
 
+export const VARIABLE_NAMESPACE = 'var'
+
+export const getVariableName = (name: string, namespace?: string) => {
+  return `${VARIABLE_NAMESPACE}-${namespace ? (namespace + '-') : ''}${name}`;
+}
+
 export function isVariableValueEqual(a: VariableValue | null | undefined, b: VariableValue | null | undefined) {
   if (a === b) {
     return true;
